2/2

+++ Blue/Assets/Scripts/BeastTamer/BeastTamer.cs
@@ -187,7 +187,7 @@ public class BeastTamer : MonoBehaviour, Boss
         }
     }
 
-    public void getDamage(int dmg) // get Damage to Boss
+    public void getDamage(float dmg) // get Damage to Boss
     {
         if (rageFlag)
         {
